회원 정보 수정 일부 기능만 구현
회원 정보 수정의 일부 기능만을 구현 하였습니다.
구현한 것
- /users에서 수정하고 싶은 회원 정보란의 수정 버튼을 클릭하면 /users/{id}/form으로 이동
- /users/{id}/form URL의 id값으로 userList에서 객체를 찾은 뒤 가져옴
- /users/{id}의 id 값에 해당하는 객체를 userList에서 찾은 뒤 입력 받아 온 새로운 정보의 객체를 List에 덮어쓰기

미구현
- 비밀번호가 일치할 시에만 수정이 가능하도록 하는 기능
- 아이디는 수정할 수 없도록 하는 기능

+++ src/main/java/com/codesquad/qna/web/QuestionController.java
@@ -37,8 +37,8 @@ public class QuestionController {
     }
 
     @GetMapping("/questions/{index}")
-    public String showQuestion(@PathVariable int index, Model model) {
-        Question selectedQuestion = questionList.get(index - 1);
+    public String showQuestion(@PathVariable long index, Model model) {
+        Question selectedQuestion = questionList.get((int)index - 1);
         model.addAttribute("question", selectedQuestion);
 
         return "qna/show";
